dispatcher-story/requirements.txt
functions-framework==3.*
google-cloud-aiplatform==1.56.0
google-cloud-firestore==2.16.0
google-cloud-tasks==2.16.3
requests==2.32.3
certifi==2024.7.4 